+++ src/main/java/com/meme/ala/core/auth/oauth/model/OAuthProvider.java
@@ -1,4 +1,4 @@
-package com.meme.ala.core.auth.oauth;
+package com.meme.ala.core.auth.oauth.model;
 
 import lombok.NoArgsConstructor;
