+++ src/main/java/com/meme/ala/core/auth/oauth/model/OAuthUserInfo.java
@@ -1,4 +1,4 @@
-package com.meme.ala.core.auth.oauth;
+package com.meme.ala.core.auth.oauth.model;
 
 public interface OAuthUserInfo {
     String getProviderId();
